Moved leaflet-initialization to utility function.
